requirements.txt
# Flask Web Framework & Security Extensions
Flask==3.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1

# Production WSGI Server (Best practice for Vercel)
gunicorn==21.2.0

# Environment Variable Management
python-dotenv==1.0.0

# API Communication & Data Handling
requests==2.31.0
fpdf2==2.7.7

# Core Dependencies (optional)
blinker==1.7.0
cffi==1.16.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
Werkzeug==3.0.1
WTForms==3.1.2